pyproject.toml
[project]
name = "ranaglyph_api"
version = "0.1.0"
description = "A secure messaging app using FastAPI, JWT authentication, and Supabase for database operations."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9"
]
dependencies = [
    "fastapi>=0.70.0",
    "uvicorn>=0.15.0",
    "supabase>=1.0.0",
    "pycryptodome>=3.10.0",
    "python-jose>=3.3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pydantic[email]>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
ranaglyph_api = ["*.env"]

[project.optional-dependencies]
test = ["pytest>=6.0"]

[project.scripts]
start-ranaglyph-api = "app.main:start"
